I can use webfont on webgl.
The key to load web-font.
1. use web-font loader.
2. use just to request just text you want to use.

If you request all text for the font, you would have wrong font at first
time. It web-font loader notify you active when the loader get first chunk of font file. If the glyph you want to use is not in first chunk of font, the first glyph display will be unexpected result.